pyproject.toml
[project]
name = "x-agent"
version = "0.1.0"
description = "Unified Python X (Twitter) agent with dual authentication"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "X Agent Contributors" }
]
dependencies = [
    "requests>=2.31.0",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.0",
    "tweepy>=4.14.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-mock>=3.11.0",
    "pytest-cov>=4.1.0",
    "responses>=0.23.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]
telemetry = [
    "opentelemetry-api>=1.26",
    "opentelemetry-sdk>=1.26",
    "opentelemetry-exporter-otlp-proto-http>=1.26",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--strict-markers --tb=short"
norecursedirs = ["_archive", ".git", "__pycache__"]

[tool.coverage.run]
source = ["src"]
omit = ["tests/*", "*/__pycache__/*"]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false

[tool.ruff]
line-length = 120
target-version = "py311"
exclude = ["_archive/", "notebooks/"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP"]
ignore = ["E501"]  # Line too long (handled by formatter)

[tool.mypy]
python_version = "3.11"
warn_return_any = false
warn_unused_configs = false
disallow_untyped_defs = false
exclude = ["_archive/"]
check_untyped_defs = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = false
# Allow untyped libraries
ignore_missing_imports = true
# Use namespace packages
explicit_package_bases = true
